Move common data and fixtures to conftest.py
