+++ tests/conftest.py
@@ -1,0 +1,27 @@
+import pytest as pytest
+
+from matching.mentee import Mentee
+from matching.mentor import Mentor
+
+
+@pytest.fixture
+def base_data() -> dict:
+    return {
+        "Your first name": "Test",
+        "Your last name": "Data",
+        "Your Civil Service email address": "test@data.com",
+        "Your job title or role": "N/A",
+        "Your department or agency": "Department of Fun",
+        "Your grade": "Grade 7",
+        "Your profession": "Policy",
+    }
+
+
+@pytest.fixture
+def base_mentee(base_data):
+    return Mentee(**base_data)
+
+
+@pytest.fixture
+def base_mentor(base_data):
+    return Mentor(**base_data)
