+++ tests/test_match.py
@@ -5,29 +5,6 @@ from matching.person import GRADES
 import pytest
 
 
-@pytest.fixture
-def base_data() -> dict:
-    return {
-        "Your first name": "Test",
-        "Your last name": "Data",
-        "Your Civil Service email address": "test@data.com",
-        "Your job title or role": "N/A",
-        "Your department or agency": "Department of Fun",
-        "Your grade": "Grade 7",
-        "Your profession": "Policy",
-    }
-
-
-@pytest.fixture
-def base_mentee(base_data):
-    return Mentee(**base_data)
-
-
-@pytest.fixture
-def base_mentor(base_data):
-    return Mentor(**base_data)
-
-
 class TestMatch:
     def test_cant_match_with_same_department(
         self, base_mentee: Mentee, base_mentor: Mentor
